Unit tests refactors & fixings
